+++ packages/storage-plugin/index.ts
@@ -1,0 +1,3 @@
+export { NgxsStoragePluginModule } from './src/storage.module';
+export { NgxsStoragePlugin } from './src/storage.plugin';
+export * from './src/symbols';