+++ packages/store/index.ts
@@ -1,0 +1,10 @@
+/**
+ * The public api for consumers of @ngxs/store
+ */
+
+export * from './src/public_api';
+
+/**
+ * The plugin api for the stuff that a plugins needs
+ */
+export * from './src/plugin_api';