+++ packages/store/src/plugin_api.ts
@@ -1,0 +1,4 @@
+export { NgxsModule } from './module';
+export { NGXS_PLUGINS, NgxsPlugin } from './symbols';
+export { StateStream } from './state-stream';
+export { getActionTypeFromInstance, setValue, getValue } from './utils';