+++ src/plugins/devtools/index.ts
@@ -1,2 +1,0 @@
-export { NgxsReduxDevtoolsPluginModule } from './devtools.module';
-export { NgxsReduxDevtoolsPlugin } from './devtools.plugin';