+++ CHANGELOG.md
@@ -1,7 +1,13 @@
-# New NPM Location
-We moved the npm package from `ngxs` to `@ngxs/store`!
+# Master
+- BREAKING: Storage plugin fixes #154
+
+# 2.0.0-rc.21 3/25/18
+- BREAKING: Plugins are now their own packages
+- Fix: Subscribe firing multiple times
+- Fix: Devtools only passing type and payload
 
 # 2.0.0-rc.20 3/23/18
+- BREAKING: We moved the npm package from `ngxs` to `@ngxs/store`!
 - Fix: devtools not sending payload #132
 - Fix: better if condition for localstorage